+++ CHANGELOG.md
@@ -37,6 +37,8 @@ to docs, or any other relevant information.
   TLS options.
 * `ClientTlsOptions` is now tagged with `#[non_exhaustive]`. Use `ClientTlsOptions::builder()` to
   construct client TLS options.
+* `ClientKeepAliveOptions` is now tagged with `#[non_exhaustive]`. Use
+  `ClientKeepAliveOptions::builder()` to construct keep-alive options.
 
 ## [0.6.0] - 2026-08-04